beater/request: move context pool
